safari-600.7-branch/LayoutTests/fast/dom/shadow/content-element-outside-shadow-expected.txt

HTMLContentElement should be able to survive even outside shadow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


<div><content></content></div>
PASS targetRenderTree is comparisonRenderTree
<div><content>Here is a text.</content></div>
PASS targetRenderTree is comparisonRenderTree
<div><content><div>Here is a </div><div>series of</div><div>blocks</div></content></div>
PASS targetRenderTree is comparisonRenderTree
<div><content><span>Here is a<span><span>series of</span><span>inlines.</span></content></div>
PASS targetRenderTree is comparisonRenderTree
<div><content>Here is a <content>nested content</content> inside content.</content></div>
PASS targetRenderTree is comparisonRenderTree
<div><content>Here is a <div>block and <content>nested content</content> inside the block</div>.</content></div>
PASS targetRenderTree is comparisonRenderTree
PASS successfullyParsed is true

TEST COMPLETE
